refactor: merge model.sympy into model
