+++ src/tensorwaves/model.py
@@ -1,4 +1,9 @@
-"""`.Function` and `.Model` implementation for `sympy`-based models."""
+"""Evaluateable physics models for amplitude analysis.
+
+The `.model` module takes care of lambdifying mathematical expressions to
+computational backends. Currently, mathematical expressions are implemented
+as `sympy` expressions only.
+"""
 
 from typing import Callable, Dict, FrozenSet, Mapping, Tuple, Union
